Update: Created a slice to save Contests, dispatching actions on a single API call from Body which is called [user],ToDo: Review the commented Code, Remove once PR merged
