+++ src/utils/appStore.js
@@ -1,8 +1,10 @@
 import { configureStore } from "@reduxjs/toolkit";
 import userReducer from "./userSlice"
+import registeredContestReducer from "./registeredContestsSlice"
 const appStore = configureStore({
     reducer : {
-      user : userReducer
+      user : userReducer,
+      registeredContests : registeredContestReducer
     }
 })
